requirements.txt
Flask>=2.0
Pillow>=9.0
gunicorn>=20.0 